Do destination check on same tick as teleport
This also helps to prevent the agent from getting stuck on belts, including the corners this time (except for the evil blue ones in most cases and the red ones from a specific angle). I want to keep some of the now unused functions from the previous fix as they might come in handy later on.