First attempt at UCS
This was tougher to integrate than I expected (I'm lying.. I was already expecting it to be a nightmare), but it works for the most part. I still need to work on the cost function a bit to make it more efficient and to include the effects of tiles and belt directions as well. Also, the phrase "refactor early, refactor often" is starting to become very relevant right now. Maybe I should even consider using metatables again..